feat: Add `hasMimeType()` and `toBlob` schema actions
# FEATURE: added `hasMimeType()` and `toBlob` schema actions

These Valibot schema actions aid with validating `datauri` strings to better ensure that fields are receiving well-formatted data from end-users.

`hasMimeType()` behaves similarly to Valibot's built-in `mimeType()` validation action, which only operates on Blobs. Without native support for Data URIs, this is necessary to ensure proper encoding of images.

`toBlob` is a transformation action that can transform a `datauri` to a Blob, which then can be used with any of Valibot's actions that operate on that data type, such as `maxSize()` to ensure that the provided value does not exceed a certain file size.