feat: Add Components, Update Application, Interactions, Add `hasSize()`
# BREAKING: reorganize types for Applications, Application Commands, Application Role Connections, Components, Interactions and Teams

Applications and Interactions got an overhaul! This necessitated reorganizing many of the files associated with both to better reflect their locations within the Discord API documentation. This may possibly result in broken imports.

# FEATURE: add Component types

Added schemas covering every Message Component type and updated the existing ones to match those in the API docs

# FEATURE: update Interaction data types and endpoints

Interactions are now up to date with their specifications within the API docs

# FEATURE: add new `hasSize()` validation action

Added a new `hasSize()` validation action which can be used with `datauri` schemas to validate approximately whether they are a given size or within a size range in terms of bytes